solver/requirements.txt
absl-py==2.1.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
immutabledict==4.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
ortools==9.11.4210
pandas==2.2.3
pika==1.3.2
protobuf==5.26.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
Werkzeug==3.0.4
